それぞれのファイルで定義していた関数を親コンポーネントに集約しました
